LPS-105616 Match REST Builder name, not Service Builder name. Type: git ls-files **BlogPosting**ExceptionMapper.java
