+++ modules/apps/headless/headless-admin-user/headless-admin-user-impl/src/main/java/com/liferay/headless/admin/user/internal/jaxrs/exception/mapper/PostalAddressZipExceptionMapper.java
@@ -14,7 +14,7 @@
 
 package com.liferay.headless.admin.user.internal.jaxrs.exception.mapper;
 
-import com.liferay.portal.kernel.exception.AddressCityException;
+import com.liferay.portal.kernel.exception.AddressZipException;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.BaseExceptionMapper;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.Problem;
 
@@ -30,17 +30,18 @@ import org.osgi.service.component.annotations.Component;
 	property = {
 		"osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Headless.Admin.User)",
 		"osgi.jaxrs.extension=true",
-		"osgi.jaxrs.name=Liferay.Headless.Admin.User.AddressCityExceptionMapper"
+		"osgi.jaxrs.name=Liferay.Headless.Admin.User.PostalAddressZipExceptionMapper"
 	},
 	service = ExceptionMapper.class
 )
-public class AddressCityExceptionMapper
-	extends BaseExceptionMapper<AddressCityException> {
+public class PostalAddressZipExceptionMapper
+	extends BaseExceptionMapper<AddressZipException> {
 
 	@Override
-	protected Problem getProblem(AddressCityException addressCityException) {
+	protected Problem getProblem(AddressZipException addressZipException) {
 		return new Problem(
-			Response.Status.BAD_REQUEST, "A city is required for an address");
+			Response.Status.BAD_REQUEST,
+			"A zip code is required for an address");
 	}
 
 }
